feat: replace otlp http & grpc with otlp proto SDK for logs
