+++ src/service/logs/mod.rs
@@ -59,8 +59,7 @@ use crate::{
 
 pub mod bulk;
 pub mod ingest;
-pub mod otlp_grpc;
-pub mod otlp_http;
+pub mod otlp;
 pub mod syslog;
 
 static BULK_OPERATORS: [&str; 3] = ["create", "index", "update"];
